81-distributed-tracing/requirements.txt
fastapi
uvicorn
requests
streamlit
opentelemetry-api
opentelemetry-sdk
opentelemetry-instrumentation-fastapi
